When replaying macro, pass shadow func
Currently when replaying recorded keys, keys pressed in insert mode are
correctly inserted into the buffer.  However keys that are repeated when
not in insert mode are ignored.  This change no longer ignores replayed
keys that are not hit in insert mode.

Fixes issue where on macro replay, f" did not move forward to the next "
character, but movement was ignored.