Attempt 2 at fixing evil-repeat
Calling `evil-repeat-stop` prevents recording the repeat information of
the current command (the repeat info is actually recorded in the
post-command-hook). This seems to be the safest way to prevent
evil-repeat from recording the first entry of the key sequence. I think
we need to ensure that the hook is run after the evil-repeat hook,
because the evil-repeat hook starts the recording.

I don't see the purpose of the insert followed by the delete, so it's
commented out, but please revert this if they serve a purpose.